:shirt: Refactor: Moves frontend code into own modules. Resolves #206.
